+++ src/main/java/frc/team555/robot/PowerUpRobot.java
@@ -1,0 +1,4 @@
+package frc.team555.robot;
+
+public class PowerUpRobot {
+}
