+++ src/main/java/frc/team555/robot/Robot.java
@@ -1,4 +1,4 @@
 package frc.team555.robot;
 
-public class Robot extends TankRobot {
+public class Robot extends frc.team555.Robot.SteamworksRobot {
 }
